+++ src/lib/packages/calendar/core/src/select/Select.ts
@@ -2,6 +2,7 @@
 import type { Calendar } from '../calendar'
 import type { Cell } from '../cell'
 import dayjs from 'dayjs'
+import { PeriodList } from '../period'
 
 interface ISelectOptions {
     allowBetweenDays?:boolean