+++ src/lib/packages/calendar/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './Timezone'
 export * from './getAllMinutesByDate'
+export * from './getDatesBetweenRange'