+++ src/lib/packages/calendar/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cluue/calendar-utils
 
+## 1.0.4
+
+### Patch Changes
+
+- Upgrate and Perf Periods
+
 ## 1.0.3
 
 ### Patch Changes